Update Django to 2.2.x
Django 2.1.x is an unsupported non LTS release that bundles jquery 3.3.1
which has a known vulnerability.

Django 2.2.13 introduced the fix:

https://github.com/django/django/commit/0daddf093f5d17a8d2ddec84ae16f6dae978a480#diff-b5669c8498215b7566b0aee2848e8e68149eddfab05a11b75fb3642e1864cb32

And Django 2.2.19 is the latest maintenance release so bumping to that.

Clean run of migrations locally ran along with test, and manual testing
in the browser with no apparent breaks.