Pinning UI image in Dockerfile
This change pins the UI image to a specific version in the Dockerfile
for the multi-stage build.

Previously, it would pull "helx-ui:latest". This made it impossible to
reliably replicate builds, and even without making appstore changes new
features would be pulled in. Now that the new UI is approaching a
releasable state, it is advantageous to pin to a specific version and,
as needed, update that dep through deliberate PR's.